Improved the way read_ast_header() handles local timestamps
Updated the read_ast_header() function to convert StartDateTimeLocal and EndDateTimeLocal to POSIXct objects with the correct time zone assigned, as long as either: (a) a whole-hour GPSUTCOffset value is available in the log file header or (b) the local time zone has been specified by the user as a function argument. As a result, a new optional argument "tz" has been added to the read_ast_header() function.